login page now handles old and new users, and there are links between the two.
