+++ api/package.json
@@ -2,6 +2,7 @@
   "dependencies": {
     "@faker-js/faker": "^9.2.0",
     "axios": "^1.7.7",
+    "bcryptjs": "^2.4.3",
     "cors": "^2.8.5",
     "dotenv": "^16.4.5",
     "express": "^4.21.1",
