--- package.json
@@ -1,5 +1,0 @@
-{
-  "dependencies": {
-    "vite": "^5.4.11"
-  }
-}
